+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add TextBuffer methods of secondary selection.
 - Fix for unicode fonts.
 - Add DisplayExt::set_highlight_data_ext() and text::StyleTableEntryExt (supports underline, strikethrough, background color...etc).
+- Add setters & getters for secondary selection, spelling and grammar colors.
 - Pull FLTK menu fixes.
 - Update FLTK (removes experimental drivers).